refactor: combining action types and actions
