+++ src/Reducers/HabitsReducer/HabitReducer.actions.ts
@@ -1,5 +1,41 @@
 import { HabitObject, Habits, HabitDate } from 'Types/Habit.types';
-import { HabitAction } from './HabitsReducer.types';
+
+interface InitAction {
+    type: 'INIT';
+    payload: Habits;
+}
+
+interface AddAction {
+    type: 'ADD';
+    payload: HabitObject;
+}
+
+interface DeleteAction {
+    type: 'DELETE';
+    payload: string;
+}
+
+interface NameAction {
+    type: 'NAME';
+    payload: string;
+    id: string;
+}
+
+interface TotalAction {
+    type: 'TOTAL';
+    payload: number;
+    id: string;
+}
+
+interface Progress {
+    type: 'PROGRESS';
+    payload: HabitDate;
+    id: string;
+    date: string;
+    complete: boolean;
+}
+
+export type HabitAction = InitAction | AddAction | DeleteAction | NameAction | TotalAction | Progress;
 
 export const habitActions = {
     init: (habits: Habits): HabitAction => ({
